refactor: create Calendar in entities
