+++ src/widgets/calendar/index.js
@@ -1,0 +1,1 @@
+export {Calendar} from './ui';
