+++ src/widgets/calendar/ui/style.css
